spec/requirements.txt
asgiref==3.5.2
concurrent-log-handler==0.9.20
coverage==6.4.4
deepmerge==1.0.1
Django==4.0.7
django-auth-ldap==4.1.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn
jira==3.4.1
ldap3==2.9.1
mock==4.0.3
mssql-django==1.1.3
pandas==1.5.2
portalocker==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.34
PyPDF2==2.10.5
python-ldap
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
PyJWT==2.3.0